Fix swapped around call note reflections and rating views
